mcp-console-host/requirements.txt
anthropic>=0.28.0
python-dotenv>=1.0.1
rich>=13.7.1
pydantic>=2.7.0
# Para modo MCP (adapter futuro)
mcp[cli]>=1.2.0
PySide6>=6.6